% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_colors.R
\docType{data}
\name{customblue}
\alias{customblue}
\title{Function to return custom colors}
\format{
An object of class \code{character} of length 1.
}
\usage{
customblue
}
\description{
Function to return custom colors
}
\keyword{datasets}
